Move fixtures to conftest.py
Test fixtures defined in conftest.py are accessible to all tests in the directory.
